fix: adjust some stuffs
