+++ client/src/types/post.d.ts
@@ -7,3 +7,23 @@ interface PostFilter {
         isTrending?: boolean;
     };
 }
+
+interface Post {
+    id: string;
+    user: {
+        id: string;
+        email: string;
+        name: string;
+        avatar: string;
+        role: string;
+    };
+    content: string;
+    medias: [{ id: string; mediaType: string; url: string }];
+    createdAt: string;
+    modifiedAt: string;
+    publishedA: string;
+    upvotedBy: string[];
+    downvotedBy: string[];
+    hashtags: [];
+    commentCount: number;
+}
